Change Project name to multus
wordcount was neither a very individual nor a catchy name so I decided
to change the name to "multus", which is latin for "many".

This means the executable is named multus, the documentation is adapted
and the preprocessor constants are prefixed with M

I decided to also rename wordcount.(c|h) as it implements mostly
project-specific code.